allow user to specify a list of HTTP methods to protect
This change allows the user to specify a delimited list of HTTP methods
that should be protected. This should be used only when the user has
concrete knowledge that all requests made via methods not in the list
are safe (i.e. do not apply an action to any data).

If the configuration property is unset then the old (protect all methods)
behavior is retained. This should be used with the understanding that
security vulnerabilities could actually be introduced by using this
feature if for example the user thinks that all actionable requests are
only available by POST requests when in fact some are available via GET
requests. If the user has excluded GET requests from the list then they
have introduced a vulnerability.