respond to sendRedirect immediately instead of delaying to the end of the filter chain
At least in Tomcat 6/7 the response object's PrintWriter stream is
closed after a call to response.sendRedirect(). This caused an
IllegalStateException to be thrown (and subsequently swallowed) and a
200 OK response with no content to get sent back to the client.

This was due to the fact that the InterceptRedirectResponse class only
kept a reference to the location of the redirect and processed the
actual redirect after the filter chain was complete. However, by this
time the wrapped response was already in a commited stated (and its
output stream closed).

This fix changes the interceptor to process the sendRedirect call
immediately. This means we can still inject the token into the redirect
URL and we don't have to delay the response.

Additionally, we only process a redirect if the HTTP GET method is
protected. If GET is unprotected there's no sense appending a token to
the URL.